backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
pydantic
python-dotenv
bcrypt
python-decouple
resend